Acabar el script main.sh

+++ iventoy.sh
@@ -4,8 +4,6 @@
 urlv="https://github.com/ventoy/PXE/releases/download/v1.0.21/iventoy-1.0.21-linux-free.tar.gz"
 urli="https://???????/mx.iso" # CAMBIAME CUANDO VAYAS A PRESENTAR
 
-
-
 wget $urlv -O iventoy.tar.gz
 
 tar xvf iventoy.tar.gz
@@ -21,6 +19,11 @@ read -p "Ahora ve a la direccion que se te indica arriba desde otro navegador. C
 echo "next-server 10.0.0.1;" >> /etc/dhcp/dhcpd.conf
 echo "filename \"iventoy_loader_16000\";" >> /etc/dhcp/dhcpd.conf # El 17000 no funciona, por mucho que lo configures.
 
-
 systemctl restart isc-dhcp-server
-wget $urli -O iso/mx.iso
+
+if [[ -f 'mx.iso' ]]
+then
+	mv ../mx.iso iso/mx.iso
+else
+	wget $urli -O iso/mx.iso
+fi
